PasswordRepository: split out PasswordSortOrder
