+++ app/src/main/java/com/zeapo/pwdstore/utils/PasswordRepository.kt
@@ -4,8 +4,6 @@
  */
 package com.zeapo.pwdstore.utils
 
-import android.content.Context
-import android.content.SharedPreferences
 import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.core.content.edit
@@ -17,7 +15,6 @@ import java.io.File
 import java.io.FileFilter
 import java.nio.file.Files
 import java.nio.file.LinkOption
-import java.util.Comparator
 import org.eclipse.jgit.api.Git
 import org.eclipse.jgit.lib.Repository
 import org.eclipse.jgit.storage.file.FileRepositoryBuilder
@@ -51,42 +48,6 @@ open class PasswordRepository protected constructor() {
 
         override fun detect(cygwinUsed: Boolean?): FS {
             return FS_POSIX_Java6_with_optional_symlinks()
-        }
-    }
-
-    enum class PasswordSortOrder(val comparator: Comparator<PasswordItem>) {
-
-        FOLDER_FIRST(Comparator { p1: PasswordItem, p2: PasswordItem ->
-            (p1.type + p1.name)
-                .compareTo(p2.type + p2.name, ignoreCase = true)
-        }),
-
-        INDEPENDENT(Comparator { p1: PasswordItem, p2: PasswordItem ->
-            p1.name.compareTo(p2.name, ignoreCase = true)
-        }),
-
-        RECENTLY_USED(Comparator { p1: PasswordItem, p2: PasswordItem ->
-            val recentHistory = Application.instance.getSharedPreferences("recent_password_history", Context.MODE_PRIVATE)
-            val timeP1 = recentHistory.getString(p1.file.absolutePath.base64())
-            val timeP2 = recentHistory.getString(p2.file.absolutePath.base64())
-            when {
-                timeP1 != null && timeP2 != null -> timeP2.compareTo(timeP1)
-                timeP1 != null && timeP2 == null -> return@Comparator -1
-                timeP1 == null && timeP2 != null -> return@Comparator 1
-                else -> p1.name.compareTo(p2.name, ignoreCase = true)
-            }
-        }),
-
-        FILE_FIRST(Comparator { p1: PasswordItem, p2: PasswordItem ->
-            (p2.type + p1.name).compareTo(p1.type + p2.name, ignoreCase = true)
-        });
-
-        companion object {
-
-            @JvmStatic
-            fun getSortOrder(settings: SharedPreferences): PasswordSortOrder {
-                return valueOf(settings.getString(PreferenceKeys.SORT_ORDER) ?: FOLDER_FIRST.name)
-            }
         }
     }
